move accumulator snapshot test to history.spec.ts
